% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean.R
\name{group_mean}
\alias{group_mean}
\title{Grouped Means}
\usage{
group_mean(data, group_var, mean_var)
}
\arguments{
\item{data}{the dataset to be used}

\item{group_var}{the grouping variable}

\item{mean_var}{the values to be used to calculate the mean, must be numeric}
}
\value{
a tibble with two columns for the group name and then the mean
}
\description{
This function uses group_by to group different variables within a column, and calculates the mean of the values for that group.
}
\examples{
data <- data.frame(
colour = rep(c("red", "blue", "yellow"), each = 5),
number = rnorm(15))
group_mean(data, colour, number)
}
